Introduces responsive_posts_archive_link
This function attempts to find the best candidate for a post archive link for display of a "View all posts" link on single posts. (Styles and exact placement TBD).

It looks first for a news template set to display "All Categories", and falls back to the posts page as set using Settings > Reading.

If no good candidate is found, it does nothing.