Add default value for sidebar_location option
In the customizer this setting defaults to 'right' as seen here:
https://github.com/bu-ist/responsive-framework/blob/develop/admin/theme-customizer.php#L294

However, when a new site is created, it does not actually have the
burf_setting_sidebar_location set until someone goes to the customizer and
save the settings. This causes get_option to return false.

This is an issue because on the responsive_body_class function this value is
used to determine the sidebar location:
https://github.com/bu-ist/responsive-framework/blob/develop/inc/extras.php#L32