Properly wrap the theme's primary stylesheet in a conditional.
WordPress supports adding IE conditionals as data attributes for enqueued styles. It does not, however, allow you to specify a conditional that targets certain versions of IE, AND other browsers (`<!--[if gt IE 8]><!-->` in our case). The extra comment tag allows the stylesheet to be loaded by non-IE browsers.

By using this filter, we can enqueue the theme's stylesheet and still load it to our needs.