Add a default value for `burf_setting_posts_sidebar_bottom`.
Previously, no default values were being set for `burf_setting_posts_sidebar_bottom` or `burf_setting_sidebar_location`. The lack of both of these options was causing the "correct behavior" (right is selected, but so is `burf_setting_posts_sidebar_bottom`, causing right to be ignored). When #248 was merged, the default of `right` correctly started to be returned for sites that had never been edited through the customizer. However, this started causing the incorrect default behavior, because `burf_setting_posts_sidebar_bottom` was considered empty due to the lack of a `true` default.