Move Customizer generated styles to the `wp_head` hook with a priority of 10.
Because the styles generated in the Customizer are not enqueued, there is no way to specify dependencies.

The wp_head hook prints all enqueued styles and scripts at priorities 8 and 9, respectively. By using priority 10 here, we can guarantee that the Customizer generate styles are output after the default stylesheet is printed.