Move narrow template tests into two separate test classes.
- One for tests that depend on constants.
- One for tests that do not depend on constants.

Tests that require constants need to be run in individual isolated processes without preserving global state to ensure the constants are not previously defined and are set to the correct value. This drastically increases the time used by each test method. Separating these test methods into two separate classes ensures that only the ones that depend on constants are run in an isolated process to prevent unnecessary slow tests methods.