Change the way that Lightgallery galleries are set up.
Currently, there is no way to load the appropriate scripts and styles for Lightgallery on a page that does not have a gallery shortcode in the post content. There are scenarios where a gallery shortcode may be output using `do_shortcode()`. In these scenarios, there needs to be a way to trigger the loading of the necessary scripts and styles.