Change the order that framework JavaScript files are loaded and allow Modernizr to be loaded in the footer.
Enqueueing Modernizr before `responsive-scripts` allows `modernizr` to be added as a dependency. This ensures Modernizr is always loaded first, incase any theme or framework JavaScript depends on Modernizr functionality.

Loading Modernizr in the footer is generally ok, but there are situations that would experience issues (see https://github.com/Modernizr/Modernizr/issues/878#issuecomment-41448059). For that reason, we will leave Modernizr in the header by default. But if a theme needs to, for some reason, they can move Modernizr to the footer using `r_modernizr_in_footer`.