Add default image size for profiles.
Currently, when profiles are displayed, the original image is loaded, even though the size of 150x150 is passed to the BU Thumbnail function. When on a profile listing page with tens or hundreds of profiles, this can be a huge performance issue if high resolution profile photos are uploaded (which we should encourage).

Profile images are also currently cropped from the center of an image out. Many images uploaded for profiles are portraits, and this can cause heads to be cut off. By changing the crop position to be center, top, we can avoid this.