"Unplug" pluggable functions in the framework.
Pluggable functions are a remnant of the earlier days in WordPress when the action and filter hook API was in its infancy. A pluggable function could be defined in a plugin or theme and it would override the WordPress core version of the function because it already exists.

The better, more modern practice is to utilize filter hooks to allow developers to change the output of a function or short circuit the function (execute code and fully override the function by returning its own result).

There are 7 functions in Responsive Framework that are "pluggable". They are listed below with the number of times they are actually overridden in our codebase and the repositories that overwrite them:

- `responsive_setup()` (0 themes override this function)
- `buniverse_shortcode()` (1 theme overrides this function: `bu-sph-responsive`. `r-sed` does also alters the output of this function, but uses the `buniverse_shortcode` filter within the function)
- `buniverse_embed_handler()` (0 themes override this function)
- `responsive_posts_widget_formats()` (1 theme overrides this function: `bu-sph-responsive`)
- `responsive_posts_navigation()` (0 themes override this function)
- `responsive_post_navigation()` (0 themes override this function)
- `responsive_post_meta()` (0 themes override this function)

This commit "unplugs" these functions so that proper filter and action hooks can be added.