Move the `clean` function to `i18n` instead of `build`.
Currently, the Grunt `clean` task is only used to empty the `languages` folder before generating a POT file. This is done because if there is a merge conflict within the POT file, then the i18n tasks will always fail. Cleaning the directory first ensures that the process will always succeed (with no other errors, of course).

If someone runs `grunt i18n`, and there is a merge conflict in the theme's POT file, the process will fail. This is because the `clean` function is only run within the `build` command. Moving it to the `i18n` command (the only place it is currently utilized) will ensure the command will succeed, even if there are merge conflicts present.