Change the default responsive gallery size to 550x550.
The one column upscaling can be fixed with CSS, but the two column is more difficult. To prevent having to use JavaScript to add a `max-width` style on each `<figure>` increasing the image size to 550x550 prevents the issue in two column galleries as long as images are at least that size.