Temporarily remove the `r_remove_news_template` filter
The `BU_News_Page_Template` class isn’t available in the admin for the filter— and adding the call to initiate the class results in a “cannot redeclare class” message. May be a namespacing problem, but for the time being: remove. This is only here so we can do theme dev locally without BU dependancies.