Bugfix for fatal error on single-site install
Added bugfix for single-site installs (not multisite) using this theme to resolve issue #3 https://github.com/bu-ist/responsive-framework/issues/3

Replaced `get_current_site()` with `network_home_url()` to prevent Fatal error: Call to undefined function when multisite is not enabled.

It appears that multisite must be enabled for `get_current__site()` to work; otherwise a fatal error occurs because the function is not available.

A possible workaround is to replace it with `network_home_url()` which returns the network’s domain and fallsback gracefully to just `home_url()` on single-site installs.

Signed-off-by: Todd Milliken <milliktr@gmail.com>