Add new template tag `responsive_the_title()`
This is intended to display the main page title for posts/pages.

This was added so that we could have a filter to prevent the output of a page title, such as the case of bu-banners wanting to use the page title instead. This gives us the benefit of not needing any `the_title` filter hooks which were being accidentally misused to prevent output, which has effects elsewhere such as nav menu titles. It is important to not use `the_title` filter because this can impact many functions that use that filter such as nav labels, get_the_title calls in widgets, etc.

This new template tag functions exactly like `the_title` and takes the same arguments. There is one additional argument for a $post->ID so something like the posts page ID can be passed to it to get the page title.

This template tag could have the potential to handle more than just single page or post titles — we can add functioanlity to detect if archive, taxonomy/term, search, or 404 queries and return the appropriate title — so that all templates run through the same page title function and can be turned off the same way in case something like bu-banners is added to a term page.