Bugfix on infinite loop
Due to an infinite loop, the code that has been isolated on the inc/calendar.php file had to be simplified to get initial values, and then based on the return of both and some additional conditions, their values could be modified on the calendar partial itself.

In the future, all data-related logic should be moved to the bu calendar plugin if possible — there is too much logic going on at the theme level which should be responsible for just templating.

I may propose a new class in the bu calendar plugin that can aid in the templating that a theme may need to do. And that the templating should be controlled by plugin shortcode so responsive-framework isn’t tied to the plugin.