Use helper functions to retrieve $boundary vars
These helper functions were created to help move logic outside of templatoing. There’s still quite a bit of logic going on depending on the current query, but that can be refactored at a later point in time — ideally moving all logic to the $buCalendar class or a new class that lives in the plugin.