Refactor proposed `responsive_the_title` function
Refactoring this so that it locks in the <h1> tag, and adds a filter for developers to change the class string.

Adds a new function called `responsive_get_the_title` which can be used to just return the expected page title for any query, without any markup.

The new `responsive_the_title` function calls this helper function, and separates the concerns of figuring out the correct page title from the markup that gets used/classnames that must get applied.