Update front page hidden page title logic
Changes the filter from `the_title` to `responsive_the_title_classes`, so that a home page that has the option checked to hide the h1 only affects this one tag, and not all instances of `the_title`.

Adds a visually hidden class rather than returning an empty string so that it gives context for accessibility.