Add filter to allow fallback font to be changed
Solves https://github.com/bu-ist/responsive-framework/issues/13 .

When a font can’t be found in `responsive_font_options`, it always returned the customizer setting or `f1` as the falllback.

This commit allows `f1` to be changed to any string. Useful in cases where the original font palette has been removed or replaced with other fonts, and a default font palette needs to be set since `f1` doesn’t exist anymore.