Upgrade to Foundation’s develop branch for updates
This is the first step in preparing to sync up with Foundation and switching to an es6 workflow.

First, upgrading to Foundation’s `develop` branch in package.json.

Second, resolve grunt error due to package.json postinstall script that runs `grunt install` and tries to pull in es6 code from Foundation before we have setup our repo to be es6-compatible. The fix is to temporarily comment out lines in Gruntfile that were bringing scripts over from Foundation via the node modules directory. When this branch is done, this step will be moot since our main js entry point will import it directly in node modules, and grunt won’t be needed to do that for us.