Move build packages to regular dependenices
This is a minor commit for semantics.

These packages are “dependencies” because we will want to include these tools in our resulting build.

Dev dependenices are for tooling and packages used to help develop the build. Dependencies are for running the code.

Since these three packages are required for Framework to run properly, they are now dependencies rather than devDependencies.