Move babel config from Gruntfile to .babelrc
This is where most babel configurations tend to be, although they could live in package.json in the browserify transform line or on its own babel line, or could live in the Gruntfile in the babelify transform.

By moving into .babelrc its easy for developers to see right away this is an es6 project before even opening any files, and that the configuration for the es6 project is visible in the root.