Add browserify shim for window.jquery for bundling
Since we will be using a window version of jquery and will not be bundling it in the app, we need to tell browserify this so it can scan imported files that use the window.jquery library and bundle them together.

Otherwise, browserify would try to include the entire jquery library in the bundle and we’d end up with WP version of jquery + our own which would be bad.