Add eslint integration, using latest WP package
This commit adds the eslint and @wordpress/eslint-plugin packages, and also autoformats the JS code in our theme according to those standards.

The WordPress eslint plugin seems to take care of a lot of rulesets, without needing to add our own rulesets (so far). As time goes on, we will most likely need to add additional rules to suppress irrelevant errors/warnings from the linter.

I got this package from https://wordpress.org/gutenberg/handbook/designers-developers/developers/packages/packages-eslint-plugin/ which seems to be the gutenberg configuration (or one of them) https://github.com/WordPress/gutenberg/tree/master/packages/eslint-plugin . There are no other eslint packages in the gutenberg plugin, so I picked this one as it seems to be the latest/greatest and worked in this case for formatting js files as I expected to see them in WP context.