Update grunt-contrib-clean, add clean:js task
This will clear out the contents of our destination js directory where compiled scripts go into, but will ignore the /vendor/ folder for now.

Useful for when entrypoint files change their filenames, so the next time the `grunt scripts` task runs it will clear out the irrelevant files and add in the new ones.

Also updated the package to 2.0.0