Overwrite nonminified compiled js files w uglified
This commit overwrites the non-minified compiled js scripts with the uglified version, since we will have source maps linked in there that point to the source code for debugging in inspector.

It is possible that for previous child themes, the non-minified regular js script file will be loaded and there won’t be a sourcemap available because it wasn’t created using this process. This isn’t a huge breaking change, just a downside on performance.

The benefit of doing this now is because the unminified versions won’t be readable anyway as bundler tools like Browserify and Webpack compile them. So rather than keep around both, it makes sense to consolidate them into one uglified file for cleaner JS files and less repo noise moving forward.