Add browserify support for lightgallery plugin
There are errors on the page when enqueuing the galleries.js file and loading it in the browser.

This is most likely due to the jQuery plugins lightgallery and lg-thumbnail — and how they don’t seem to work with CommonJS modules because of how their package.json “main” key resolves to an entry point that isn’t a CommonJS module.

To get around this, we add the Browserify “browser” key to specify what the entry point should be, and what we’d like to name that when we do require() or import ‘’;.

By doing this, we can now use es6 imports as `import ‘lightgallery`, and that will bundle the dependency with other others that share the lightgallery functionality.