Restore copying lightgallery assets to js/vendor
Ideally there is an alternative lightbox/gallery plugin that is not jQuery based and does not require hardcoded font files/images in the CSS rules that can’t be changed.

However, to deal with it in the interim, we will copy just the fonts, images, and css files needed for lightgallery js functionality; but we will still leverage es6 imports for the javascript functionality. This will allow a future update of lightgallery to be easier to maintain if they remove dependencies for their css/fonts.