Fix warnings on responnsie_get_posts_archive_link
Two warnings were being thrown on this function, if it was called on a page that didn’t have any category terms assigned to it.

- `Invalid argument supplied for foreach()`: due to wp_list_pluck() usage at wp-content/themes/responsive-framework-2-x/inc/template-tags.php:804
- `in_array() expects parameter 2 to be array, boolean given`: due to in_array() usage at wp-content/themes/responsive-framework-2-x/inc/template-tags.php:817

This fix sets $post_cat_ids to an empty array to resolve the second warning. And then it fills the array if it has terms to pluck term IDs from to resolve the first warning.