Add action hook support to 404/simplify template
- Adds the action hooks that were added in previous commit (see page.php), so that the new page title placement will fire on the 404 template.
- Replaces `<div class="error-404 not-found content-area">` with `<article class="error-404 not-found content-area">`, so that way the article action hooks make sense.
- Removes the redundant <header> wrapper, so its more consistent with other templates.
- Removes the redundant `<div class=“page-content”>` wrapper, so its more consistent with other templates.

There were no styling side-effects from making these changes to the 404 template in Responsive Framework out-of-the-box. This commit simplifies the template so it is more consistent with page.php, and in the future could potentially be consolidated into index.php. Perhaps the content inside could be in a partial like `template-parts/content-404.php`.