Add action hook support to single.php
Adds the action hooks that were added in previous commits (see page.php), so that the new page title placement will fire on the single.php template.

Also removed the redundant `<header>` tag that was wrapping the title. It doesn’t seem to be used for anything out-of-the-box; there is no visual difference between having this tag wrap the title, or removing it. Since it doesn’t even have a class and isn’t offering much and is making page titles inconsistent from the rest of the theme, it has been removed so it follows the standard title placement and markup.