Simplify placement of title by removing wp hook
The original intention of responsive_the_title_location() was to fire the title in different hooks if we needed to.

Since the title is only being added in one place in the entire theme, this function can be reduced to a one liner action hook that fires `responsive_the_title` after the opening article tag, using the default priority of `10`.