PHP Notice: Undefined index -1
When there were no events to display, line 61 was throwing a notice because it was attempting to retrieve an index from an empty array.

Moved this line closer to where its used in line 66, where we know there are events (there are actually 25 or more). This check would work in that case.