Remove default value for customizer setting
On both responsi 1x + 2x themes, if you go to Customizer > Content Options > uncheck the setting for “Keep the posts sidebar on bottom”, and refresh the page — the setting will not stick.

This is because it was unchecked on save, so it doesn’t exist in the database anymore, and on the next page load it defaults to true because of this line.

By removing the default value, it will stay as a false value on the next page load.