:bug: Add pagination to search template
The WP search template did not have pagination, so if used it would only show the latest x posts (based on the site’s default setting for posts_per_page).

This commit adds the pagination back in utilizing the template tag already available in responsi for outputting a current query’s pagination.

Also added a conditional to the pagination function to change the $archive_type to ‘search’ to avoid seeing ‘Older’ and ‘Newer’ in the pagination, and instead use the default which is ‘Previous’ and ‘Next’.