add responsive_html_class() function and filter hook
This new function will let themes easily add a class to the HTML tag. For instance if a template or component needs a class name on the HTML tag in addition to the body tag for styling purposes.

This is needed to better support the Sequential Cards Gutenberg Block in BU Prepress and avoid using JS to add the class to the `<html>` tag on posts with the block present.